Corrected Obj loader, now imports texture coordinates, normals and vertices properly. Corrected texture and light mixing in the shader. Corrected camera translation, to stop camera from rotating. Merge commit between texture implementation and object loading.

+++ GameEngine/Source/GeometricObject.h
@@ -15,7 +15,7 @@ protected:
 	Material MaterialColors;
 	std::vector <Texcoord> Texcoords;
 	std::vector <Normal> Normals;
-	std::vector <GLubyte> Indices;
+	std::vector <GLuint> Indices;
 	int indicesCount;
 	int verticesCount;
 	Matrix4f transformations;
@@ -28,7 +28,7 @@ protected:
 public:
 
 	GeometricObject(BufferObjects* buffer, Scene* scene, Mesh mesh);
-	GeometricObject(BufferObjects* buffer, Scene* scene, Vertex* Vertices, int verticesSize, GLubyte* Indices, int indicesSize);
+	GeometricObject(BufferObjects* buffer, Scene* scene, Vertex* Vertices, int verticesSize, GLuint* Indices, int indicesSize);
 	void draw(Matrix4f parentNodeTransformations, Texture* texture);
 	void updateBuffer();
 	void translate(Vector3f translation);
